requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
scikit-learn
psycopg2-binary
supabase
python-dotenv
prometheus-client
pandas
torch==2.11.0
torchvision==0.26.0
open_clip_torch==3.3.0
ftfy==6.3.1
timm==1.0.26
pillow
python-multipart